refactor: token 픽스쳐 수정
