+++ backend/sokdak/src/test/java/com/wooteco/sokdak/comment/acceptance/CommentAcceptanceTest.java
@@ -6,7 +6,7 @@ import static com.wooteco.sokdak.util.fixture.HttpMethodFixture.httpDeleteWithAu
 import static com.wooteco.sokdak.util.fixture.HttpMethodFixture.httpGet;
 import static com.wooteco.sokdak.util.fixture.HttpMethodFixture.httpPost;
 import static com.wooteco.sokdak.util.fixture.HttpMethodFixture.httpPostWithAuthorization;
-import static com.wooteco.sokdak.util.fixture.MemberFixture.getTokensForReport;
+import static com.wooteco.sokdak.util.fixture.MemberFixture.getFiveTokens;
 import static com.wooteco.sokdak.util.fixture.PostFixture.addNewPost;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
@@ -116,7 +116,7 @@ class CommentAcceptanceTest extends AcceptanceTest {
         Long postId = addNewPost();
         Long commentId = addCommentAndGetCommentId(postId);
         addCommentAndGetCommentId(postId);
-        List<String> tokens = getTokensForReport();
+        List<String> tokens = getFiveTokens();
 
         for (int i = 0; i < 5; ++i) {
             ReportRequest reportRequest = new ReportRequest("댓글신고");
