+++ backend/sokdak/src/test/java/com/wooteco/sokdak/member/acceptance/MemberAcceptanceTest.java
@@ -1,6 +1,6 @@
 package com.wooteco.sokdak.member.acceptance;
 
-import static com.wooteco.sokdak.util.fixture.HttpMethodFixture.getChrisToken;
+import static com.wooteco.sokdak.util.fixture.MemberFixture.getChrisToken;
 import static com.wooteco.sokdak.util.fixture.HttpMethodFixture.httpGetWithAuthorization;
 import static com.wooteco.sokdak.util.fixture.HttpMethodFixture.httpPatchWithAuthorization;
 import static org.assertj.core.api.Assertions.assertThat;
