+++ backend/sokdak/src/test/java/com/wooteco/sokdak/util/fixture/CommentFixture.java
@@ -1,6 +1,6 @@
 package com.wooteco.sokdak.util.fixture;
 
-import static com.wooteco.sokdak.util.fixture.HttpMethodFixture.getChrisToken;
+import static com.wooteco.sokdak.util.fixture.MemberFixture.getChrisToken;
 import static com.wooteco.sokdak.util.fixture.HttpMethodFixture.httpPostWithAuthorization;
 
 import com.wooteco.sokdak.comment.dto.NewCommentRequest;
