+++ backend/sokdak/src/test/java/com/wooteco/sokdak/util/fixture/HttpMethodFixture.java
@@ -94,19 +94,4 @@ public class HttpMethodFixture {
     public static String getExceptionMessage(ExtractableResponse<Response> response) {
         return response.jsonPath().getString("message");
     }
-
-    public static String getChrisToken() {
-        LoginRequest loginRequest = new LoginRequest("chris", "Abcd123!@");
-        return httpPost(loginRequest, "/login").header(AUTHORIZATION);
-    }
-
-    public static String getJoshToken() {
-        LoginRequest loginRequest = new LoginRequest("josh", "Abcd123!@");
-        return httpPost(loginRequest, "/login").header(AUTHORIZATION);
-    }
-
-    public static String getAdminToken() {
-        LoginRequest loginRequest = new LoginRequest("admin1", "Abcd123!@");
-        return httpPost(loginRequest, "/login").header(AUTHORIZATION);
-    }
 }
